Currency exchange for deposits implemented

+++ Backend/Database/Models/Transaction.cs
@@ -7,9 +7,9 @@ namespace VaultX.Database.Models
         public uint DestinationAccountId { get; set; }
         public string Depositant { get; set; }
         public string CurrencyCode { get; set; }
-        public decimal ConversionRate { get; set; }
-        public decimal OriginalAmount { get; set; }
-        public decimal ConvertedAmount { get; set; }
+        public double ConversionRate { get; set; }
+        public double OriginalAmount { get; set; }
+        public double ConvertedAmount { get; set; }
         public DateTime Timestamp { get; set; }
         public Account DestinationAccount { get; set; }
     }
@@ -19,9 +19,9 @@ namespace VaultX.Database.Models
         public uint Id { get; set; }
         public uint OriginAccountId { get; set; }
         public uint DestinationAccountId { get; set; }
-        public decimal ConversionRate { get; set; }
-        public decimal AmountInOriginCurrency { get; set; }
-        public decimal AmountInDestinationCurrency { get; set; }
+        public double ConversionRate { get; set; }
+        public double AmountInOriginCurrency { get; set; }
+        public double AmountInDestinationCurrency { get; set; }
         public DateTime Timestamp { get; set; }
         public Account OriginAccount { get; set; }
         public Account DestinationAccount { get; set; }
